// proto/gen/ts/auth/v2/auth.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.0
//   protoc               unknown
// source: auth/v2/auth.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "auth.v2";

export enum Country {
  COUNTRY_UNSPECIFIED = 0,
  COUNTRY_EN = 1,
  COUNTRY_IT = 2,
  COUNTRY_FR = 3,
  COUNTRY_DE = 4,
  COUNTRY_RU = 5,
  UNRECOGNIZED = -1,
}

export interface User {
  id: string;
  email: string;
  password: string;
  firstname: string;
  lastname: string;
  nickname: string;
  country: Country;
}

export interface AuthRegisterRequest {
  email: string;
  password: string;
  firstname: string;
  lastname: string;
  nickname: string;
  country: Country;
}

export interface AuthRegisterResponse {
  accessToken: string;
  refreshToken: string;
  user: User | undefined;
}

export interface AuthLoginRequest {
  email: string;
  password: string;
}

export interface AuthLoginResponse {
  accessToken: string;
  refreshToken: string;
  user: User | undefined;
}

export interface AuthRefreshRequest {
  refreshToken: string;
}

export interface AuthRefreshResponse {
  accessToken: string;
  refreshToken: string;
}

export interface AuthLogoutRequest {
  refreshToken: string;
}

export interface AuthLogoutResponse {
  message: string;
}

export const AUTH_V2_PACKAGE_NAME = "auth.v2";

export interface AuthServiceClient {
  register(request: AuthRegisterRequest): Observable<AuthRegisterResponse>;

  login(request: AuthLoginRequest): Observable<AuthLoginResponse>;

  refreshToken(request: AuthRefreshRequest): Observable<AuthRefreshResponse>;

  logout(request: AuthLogoutRequest): Observable<AuthLogoutResponse>;
}

export interface AuthServiceController {
  register(
    request: AuthRegisterRequest,
  ): Promise<AuthRegisterResponse> | Observable<AuthRegisterResponse> | AuthRegisterResponse;

  login(request: AuthLoginRequest): Promise<AuthLoginResponse> | Observable<AuthLoginResponse> | AuthLoginResponse;

  refreshToken(
    request: AuthRefreshRequest,
  ): Promise<AuthRefreshResponse> | Observable<AuthRefreshResponse> | AuthRefreshResponse;

  logout(request: AuthLogoutRequest): Promise<AuthLogoutResponse> | Observable<AuthLogoutResponse> | AuthLogoutResponse;
}

export function AuthServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["register", "login", "refreshToken", "logout"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("AuthService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("AuthService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const AUTH_SERVICE_NAME = "AuthService";